// data-parser/src/app/models/categories.model.ts
export enum Categories {
    INDOOR_LAMPS = "lampy-wewnetrzne",
    OUTDOOR_LAMPS = "lampy-zewnetrzne-i-ogrodowe",
}

export enum SubCategoriesIndoorLamps {
    ZYRANDOLE = "twoje-lampy-zyrandole",
    WIESZACE = "twoje-lampy-oswietlenie-wiszace",
    BIURKOWE = "lampy-biurkowe",
}

export enum SubCategoriesOutdoorLamps {
    DEKORACYJNE = "lampy-dekoracyjne-ogrodowe",
    SOLARNE = "lampy-ogrodowe-solarne",
    SUFITOWE = "lampy-zewnetrzne-sufitowe",
}